+ IgniteClusterSnapshotDeleteTest
